+++ src/v2019C1/extensions/presentation/index.ts
@@ -1,0 +1,5 @@
+import * as presentationQueries from './query'
+
+export const presentation = {
+	query: presentationQueries,
+}