+++ src/v2019C1/extensions/presentation/query/index.ts
@@ -1,0 +1,1 @@
+export { extractElementTitle } from './extract-element-title'